+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - [spiral/validation] Add array::count, array::range, array::shorter and array::longer rules (#435)
   - [spiral/cache] New component with common interfaces (RR2.0 support)
   - [spiral/queue] New component with common interfaces (RR2.0 support)
+  - [spiral/views] [Allow custom loader in ViewManager](https://github.com/spiral/framework/issues/488)
 
 ## v2.8.0 - 2021-06-03